Add files via upload
This the whole final e-skin dissertation, it's working principle just like a keyboard. Quite simple, But it is already widely used in our daily work, and free time. Key board can only detect one location to verify whether it be pressed or not. We can consider it as 0/1 in digital electronic field. 
But this project that I finished is try to solve whether we can detect the very small pressing or manipulation signals and reflect it's real force parameter. we also continue the similar structure, but use a very linear respond material, which means 1/R and force are linear relationship. 
If you are familiar with Amplifier units or normal circuit in electronic information field, the above relationship can be easily transformed into a equal relationship between force and readout-number.